requirements.txt
# Investment Radar - Requirements
# Python 3.11+

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0

# Scheduling
apscheduler>=3.10.0

# Database
sqlalchemy>=2.0.0

# LLM
openai>=1.0.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
